more header file re-org
