+++ tests/core/game_board_concept_test.cpp
@@ -1,4 +1,4 @@
-#include <builders.hpp>
+#include <boardstate/zobrist_factories.hpp>
 #include <concepts/board_state_calculator_registry.hpp>
 #include <concepts/composite_concepts.hpp>
 #include <concepts/space_info_provider.hpp>
