+++ tests/core/move_evaluator_minimax_for_concepts/move_evaluator_minimax_for_concepts_test.cpp
@@ -1,7 +1,8 @@
-#include <builders.hpp>
+#include <boardstate/zobrist_factories.hpp>
 #include <concepts/move_evaluator.hpp>
 #include <gameboard/game_board_for_concepts.hpp>
 #include <gtest/gtest.h>
+#include <moveselection/minimax_evaluator_factory.hpp>
 #include <moveselection/move_evaluator_minimax_for_concepts.hpp>
 #include <piecepoints/piece_position_points_for_concepts.hpp>
 #include <type_traits>
