+++ tests/core/piece_position_points_concept_test.cpp
@@ -1,4 +1,4 @@
-#include <builders.hpp>
+#include <boardstate/zobrist_factories.hpp>
 #include <concepts/piece_value_provider.hpp>
 #include <gtest/gtest.h>
 #include <piecepoints/base_position_offset.hpp>
