+++ tests/core/zobrist_component_concept_test.cpp
@@ -1,6 +1,6 @@
 #include <array>
 #include <board_map_fixture.hpp>
-#include <builders.hpp>
+#include <boardstate/zobrist_factories.hpp>
 #include <concepts/multi_board_state_provider.hpp>
 #include <concepts/single_board_state_provider.hpp>
 #include <gameboard/game_board_for_concepts.hpp>
