+++ tests/core/zobrist_coordinator_concept_test.cpp
@@ -1,4 +1,4 @@
-#include <builders.hpp>
+#include <boardstate/zobrist_factories.hpp>
 #include <concepts/board_state_coordinator.hpp>
 #include <concepts/single_board_state_provider.hpp>
 #include <concepts>
